chore: webpack config refactor
