+++ packages/account-v2/src/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { makeLazyLoader, moduleLoader } from './utils';
+import { makeLazyLoader, moduleLoader } from './utils/loader';
 
 const LazyApp = makeLazyLoader(
     () => moduleLoader(() => import(/* webpackChunkName: "account-v2", webpackPreload: true */ './App')),
