+++ packages/account/package.json
@@ -35,7 +35,6 @@
         "@deriv/integration": "1.0.0",
         "@deriv/shared": "^1.0.0",
         "@deriv/stores": "^1.0.0",
-        "@deriv/account-v2": "2.0.0",
         "@deriv/translations": "^1.0.0",
         "bowser": "^2.9.0",
         "classnames": "^2.2.6",
